Move `BlockIdentifier` & other structs and impls to their own crate
