+++ crates/rpc/eth/block.rs
@@ -1,13 +1,11 @@
-use std::fmt::Display;
-
 use ethereum_rust_chain::find_parent_header;
-use serde::Deserialize;
 use serde_json::Value;
 use tracing::info;
 
 use crate::{
     types::{
         block::RpcBlock,
+        block_identifier::{BlockIdentifier, BlockIdentifierOrHash},
         receipt::{RpcReceipt, RpcReceiptBlockInfo, RpcReceiptTxInfo},
     },
     utils::RpcErr,
@@ -17,9 +15,7 @@ use ethereum_rust_core::{
     rlp::encode::RLPEncode,
     types::{calculate_base_fee_per_blob_gas, BlockBody, BlockHash, BlockHeader, BlockNumber},
 };
-use ethereum_rust_storage::{error::StoreError, Store};
-
-use super::account::BlockIdentifierOrHash;
+use ethereum_rust_storage::Store;
 
 pub struct GetBlockByNumberRequest {
     pub block: BlockIdentifier,
@@ -42,65 +38,6 @@ pub struct GetBlockReceiptsRequest {
 #[derive(Clone, Debug)]
 pub struct GetRawHeaderRequest {
     pub block: BlockIdentifier,
-}
-
-#[derive(Clone, Debug)]
-pub enum BlockIdentifier {
-    Number(BlockNumber),
-    Tag(BlockTag),
-}
-
-#[derive(Deserialize, Default, Clone, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum BlockTag {
-    Earliest,
-    Finalized,
-    Safe,
-    #[default]
-    Latest,
-    Pending,
-}
-
-impl BlockIdentifier {
-    pub fn resolve_block_number(&self, storage: &Store) -> Result<Option<BlockNumber>, StoreError> {
-        match self {
-            BlockIdentifier::Number(num) => Ok(Some(*num)),
-            BlockIdentifier::Tag(tag) => match tag {
-                BlockTag::Earliest => storage.get_earliest_block_number(),
-                BlockTag::Finalized => storage.get_finalized_block_number(),
-                BlockTag::Safe => storage.get_safe_block_number(),
-                BlockTag::Latest => storage.get_latest_block_number(),
-                BlockTag::Pending => storage.get_pending_block_number(),
-            },
-        }
-    }
-
-    pub fn parse(serde_value: Value, arg_index: u64) -> Result<Self, RpcErr> {
-        // Check if it is a BlockTag
-        if let Ok(tag) = serde_json::from_value::<BlockTag>(serde_value.clone()) {
-            return Ok(BlockIdentifier::Tag(tag));
-        };
-        // Parse BlockNumber
-        let Ok(hex_str) = serde_json::from_value::<String>(serde_value) else {
-            return Err(RpcErr::BadParams);
-        };
-        // Check that the BlockNumber is 0x prefixed
-        let Some(hex_str) = hex_str.strip_prefix("0x") else {
-            return Err(RpcErr::BadHexFormat(arg_index));
-        };
-        // Parse hex string
-        let Ok(block_number) = u64::from_str_radix(hex_str, 16) else {
-            return Err(RpcErr::BadHexFormat(arg_index));
-        };
-        Ok(BlockIdentifier::Number(block_number))
-    }
-
-    pub fn resolve_block_header(&self, storage: &Store) -> Result<Option<BlockHeader>, StoreError> {
-        match self.resolve_block_number(storage)? {
-            Some(block_number) => storage.get_block_header(block_number),
-            _ => Ok(None),
-        }
-    }
 }
 
 impl RpcHandler for GetBlockByNumberRequest {
@@ -322,26 +259,5 @@ pub fn get_blob_base_fee(storage: &Store) -> Result<Value, RpcErr> {
             serde_json::to_value(format!("{:#x}", blob_base_fee)).map_err(|_| RpcErr::Internal)
         }
         _ => Err(RpcErr::Internal),
-    }
-}
-
-impl Display for BlockIdentifier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BlockIdentifier::Number(num) => num.fmt(f),
-            BlockIdentifier::Tag(tag) => match tag {
-                BlockTag::Earliest => "Earliest".fmt(f),
-                BlockTag::Finalized => "Finalized".fmt(f),
-                BlockTag::Safe => "Safe".fmt(f),
-                BlockTag::Latest => "Latest".fmt(f),
-                BlockTag::Pending => "Pending".fmt(f),
-            },
-        }
-    }
-}
-
-impl Default for BlockIdentifier {
-    fn default() -> BlockIdentifier {
-        BlockIdentifier::Tag(BlockTag::default())
     }
 }
